requirements.txt
# Web framework
Flask==2.3.2

# Cross-Origin Resource Sharing
flask-cors==5.0.1

# WSGI server
gunicorn==20.1.0

# MySQL driver
mysql-connector-python==8.1.0

# SNMP libraries
pysnmp==4.4.12
pyasn1>=0.5.0,<0.6.0
pyasn1-modules==0.3.0

# Scheduler
APScheduler==3.10.1

# HTTP clients & testing
requests
pytest>=7.0
httpx>=0.25
pytest-asyncio>=0.21

# FastAPI / Uvicorn (some tests rely on FastAPI)
fastapi>=0.95
uvicorn[standard]>=0.22

# Redis client (fixes ModuleNotFoundError seen in logs)
redis>=4.5.0
